Controller and tests refactored, Request validation implemented
