+++ TaskOrganizer/Gateway/TaskOrganizer.Api/Models/TaskModel.cs
@@ -1,4 +1,5 @@
 using System;
+using TaskOrganizer.Api.Validation;
 
 namespace TaskOrganizer.Api.Models
 {
@@ -12,5 +13,16 @@ namespace TaskOrganizer.Api.Models
         public DateTime EstimatedDate { get; set; }
         public DateTime? StartDate { get; set;}
         public DateTime? EndDate { get; set;}
+
+        public void IsValid()
+        {
+            var validation = new TaskModelValidator();
+            
+            if(this.Progress.Equals("ToDo"))
+                validation.ValidateToDo(this);
+            
+            if(this.Progress.Equals("InProgress"))
+                validation.ValidateInProgress(this, nameof(EndDate));
+        }
     }
 }
