+++ TaskOrganizer/Gateway/TaskOrganizer.Api/TaskOrganizer.Api.csproj
@@ -8,4 +8,8 @@
     <ProjectReference Include="..\TaskOrganizer.Repository\TaskOrganizer.Repository.csproj" />
 </ItemGroup>
 
+<ItemGroup>
+  <PackageReference Include="FluentValidation" Version="8.6.2" />
+</ItemGroup>
+
 </Project>
